Improve: 2x wider GEMM steps on LASX

+++ include/numkong/spatials/loongsonasx.h
@@ -102,12 +102,12 @@ nk_define_cross_normalized_symmetric_(euclidean, u8, loongsonasx, u8, u32, /*nor
                                       nk_partial_load_b32x4_serial_, nk_store_b128_loongsonasx_,
                                       nk_partial_store_b32x4_serial_, 1)
 
-nk_define_cross_normalized_packed_(angular, bf16, loongsonasx, bf16, f32, f32, /*norm_value_type=*/f32, f32,
+nk_define_cross_normalized_packed_(angular, bf16, loongsonasx, bf16, bf16, f32, /*norm_value_type=*/f32, f32,
                                    nk_b128_vec_t, nk_dots_packed_bf16_loongsonasx,
                                    nk_angular_through_f32_from_dot_loongsonasx_, nk_dots_reduce_sumsq_bf16_,
                                    nk_load_b128_loongsonasx_, nk_partial_load_b32x4_serial_, nk_store_b128_loongsonasx_,
                                    nk_partial_store_b32x4_serial_, 1)
-nk_define_cross_normalized_packed_(euclidean, bf16, loongsonasx, bf16, f32, f32, /*norm_value_type=*/f32, f32,
+nk_define_cross_normalized_packed_(euclidean, bf16, loongsonasx, bf16, bf16, f32, /*norm_value_type=*/f32, f32,
                                    nk_b128_vec_t, nk_dots_packed_bf16_loongsonasx,
                                    nk_euclidean_through_f32_from_dot_loongsonasx_, nk_dots_reduce_sumsq_bf16_,
                                    nk_load_b128_loongsonasx_, nk_partial_load_b32x4_serial_, nk_store_b128_loongsonasx_,
